Replaced class files. Converted coordinates from double to int

+++ src/GeneratorDanych/BreweryFactory.java
@@ -1,10 +1,12 @@
 package GeneratorDanych;
 
+import SiecPrzeplywowa.Brewery;
+
 import java.util.Random;
 
 class BreweryFactory implements IFactory<Brewery> {
     @Override
-    public Brewery Create(Random random, double x, double y) {
+    public Brewery Create(Random random, int x, int y) {
         double prodCap = random.nextDouble(10, 10000);
         return new Brewery(x, y, prodCap);
     }
