+++ src/GeneratorDanych/IFactory.java
@@ -3,5 +3,5 @@ package GeneratorDanych;
 import java.util.Random;
 
 interface IFactory<T> {
-    T Create(Random random, double x, double y);
+    T Create(Random random, int x, int y);
 }
